+++ packages/std/src/math/fraction.rs
@@ -10,6 +10,7 @@ pub trait Fraction<T>: Sized {
     /// Returns the multiplicative inverse `q/p` for fraction `p/q`.
     ///
     /// If `p` is zero, None is returned.
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     fn inv(&self) -> Option<Self>;
 }
 
@@ -58,6 +59,7 @@ macro_rules! impl_mul_fraction {
             }
 
             /// Same operation as `checked_mul_floor` except unwrapped
+            #[must_use = "this returns the result of the operation, without modifying the original"]
             pub fn mul_floor<F: Fraction<T>, T: Into<$Uint>>(self, rhs: F) -> Self {
                 self.checked_mul_floor(rhs).unwrap()
             }
@@ -89,6 +91,7 @@ macro_rules! impl_mul_fraction {
             }
 
             /// Same operation as `checked_mul_ceil` except unwrapped
+            #[must_use = "this returns the result of the operation, without modifying the original"]
             pub fn mul_ceil<F: Fraction<T>, T: Into<$Uint>>(self, rhs: F) -> Self {
                 self.checked_mul_ceil(rhs).unwrap()
             }
@@ -119,6 +122,7 @@ macro_rules! impl_mul_fraction {
             }
 
             /// Same operation as `checked_div_floor` except unwrapped
+            #[must_use = "this returns the result of the operation, without modifying the original"]
             pub fn div_floor<F: Fraction<T>, T: Into<$Uint>>(self, rhs: F) -> Self
             where
                 Self: Sized,
@@ -156,6 +160,7 @@ macro_rules! impl_mul_fraction {
             }
 
             /// Same operation as `checked_div_ceil` except unwrapped
+            #[must_use = "this returns the result of the operation, without modifying the original"]
             pub fn div_ceil<F: Fraction<T>, T: Into<$Uint>>(self, rhs: F) -> Self
             where
                 Self: Sized,